Add files via upload
We worked on a 3.5 GB large data set, as a team. We progressed by following the data
science methodology. We performed our analysis with comprehensive and statistical validity. The
project was divided into 2 purposes. First goal; We made text classification with deep learning.
Second goal; this path forward was our main road we planned. In this path, we proceeded by
comparing the processes until the model installation with the alternatives. We established our artificial
learning models. We tested and compared. Our results became meaningful and consistent with the
motivation of the project. Our project was among the interesting and successful projects. In the
bootcamp graduation ceremony in which various projects have been presented as well.